Standardize math types (see #94). API breaking changes. Clean up type warnings.
- Matrix4F objects are now defined to be column-major, and a matrix `M` now
transforms a vector `v` by multiplying with the column vector on the right, like
`Mv`. This shouldn't impact your code unless you were using `GetRowVec` or `GetF`

- `sample_utils`' `vec.h`, `vec.cpp`, `camera.h`, and `camera.cpp` have been removed.
These were copies of their corresponding GVDB classes that were slowly diverging
over time, and are not needed (the only sample impacted as a result is `gFluidSurface`).
If you were using these, include the GVDB headers for drop-in replacements.

- `Vector3DI` and `Vector3DF` are now templatized instantiations of a single
`Vector3D` class. To accomodate this, its functions have been moved inline to
`gvdb_vec.h`. This might increase compilation times a bit - sorry about that!

- `Matrix4F::operator* (const Vector3DF &op)` now returns a `Vector3DF`, the result
of transforming the vector by the matrix, instead of a `Matrix4F`, the result of
multiplying the matrix by `diag(op.x, op.y, op.z, 1)`. The old behavior has been
renamed to `Matrix4F::ScaleInPlace`.

- `ActivateSpaceAtLevel` and `isActive` now take a `Vector3DI` pos instead of a
`Vector3DF`. Previously, the `Vector3DF` was internally casted to a `Vector3DI`.

- `getNumNodes`, `getNumUsedNodes`, and `getNumTotalNodes` now return 64-bit
unsigned integers in order to represent the underlying `DataPtr` structure more
accurately. However, most users of these functions cast them to `ints` in any case.

- The order of arguments in `mmult` have been swapped.

- Removes unused variables `camivprow0...camivprow3` from `ScnInfo`.

- `AllocateNeighbors` now takes a `uint64` instead of an `int`. However, this
change is not very impactful.

- Renames the variables in `Matrix4F::operator()`. By coincidence, this produces
exactly the same behavior as before, so no code changes are needed.

Also includes changes throughout the codebase to fix the type casting warnings
VS 2019 was warning about, removes the `GVDB_VEC`... variables, and adds
documentation for math class functions.

Note that you may have to do a clean rebuild of the project after pulling this
commit, as the mangled type names for the `Vector3` classes have changed.

+++ source/g3DPrint/main_3dprint.cpp
@@ -116,10 +116,10 @@ void Sample::revoxelize()
 	// allows the application to setup rendering to convert from unit voxel grid to world coordinates.
 	// To apply voxel size for 3D printing, we include it in the xform matrix for SolidVoxelize.
 	switch (m_voxelsize_select) {
-	case 0:	m_voxel_size = 0.5;		break;
-	case 1:	m_voxel_size = 0.4;		break;
-	case 2:	m_voxel_size = 0.3;		break;
-	case 3:	m_voxel_size = 0.2;		break;
+	case 0:	m_voxel_size = 0.5f; break;
+	case 1:	m_voxel_size = 0.4f; break;
+	case 2:	m_voxel_size = 0.3f; break;
+	case 3:	m_voxel_size = 0.2f; break;
 	};
 
 	// Create a transform	
@@ -283,14 +283,15 @@ void Sample::render_section ()
 	float h = (float) getHeight();
 	Vector3DF world;
 	world = m_pivot * m_part_size / m_voxel_size;	// GVDB voxel grid to world coordinates
-	world.y *= 1.0 - getCurY() / h;					// Select cross-section on world y-axis 
+	world.y *= 1.0f - getCurY() / h;				// Select cross-section on world y-axis 
 	gvdb1.getScene()->SetCrossSection ( world, Vector3DF(world.x, 1.f, world.z) );
 
 	gvdb1.Render ( SHADE_SECTION2D, 0, 1 );		
 
 	gvdb1.ReadRenderTexGL ( 1, gl_section_tex );
 	
-	renderScreenQuadGL ( gl_section_tex, -1, 0, 0, getWidth()/4, getHeight()/4, 0 );
+	renderScreenQuadGL ( gl_section_tex, -1, 0, 0,
+		static_cast<float>(getWidth())/4, static_cast<float>(getHeight())/4, 0 );
 }
 
 void Sample::display()
@@ -357,7 +358,7 @@ void Sample::draw_topology ( VolumeGVDB* gvdb )
 	Vector3DF bmin, bmax;
 	Node* node;
 	for (int lev=0; lev < 5; lev++ ) {				// draw all levels
-		int node_cnt = gvdb->getNumNodes(lev);
+		int node_cnt = static_cast<int>(gvdb->getNumNodes(lev));
 		for (int n=0; n < node_cnt; n++) {			// draw all nodes at this level
 			node = gvdb->getNodeAtLevel ( n, lev );
 			bmin = gvdb->getWorldMin ( node ) * vs; // get node bounding box
